+++ src/renderer/src/components/icons/ModelIcon.vue
@@ -68,12 +68,14 @@ import zenmuxColorIcon from '@/assets/llm-icons/zenmux-color.svg?url'
 import burncloudColorIcon from '@/assets/llm-icons/burncloud-color.svg?url'
 import xiaomiColorIcon from '@/assets/llm-icons/xiaomi.png?url'
 import o3fanColorIcon from '@/assets/llm-icons/o3-fan.png?url'
+import opencodeIcon from '@/assets/llm-icons/opencode.png?url'
 
 // 导入所有图标
 const icons = {
   'kimi-cli': moonshotColorIcon,
   'claude-code-acp': claudeColorIcon,
   'codex-acp': openaiColorIcon,
+  opencode: opencodeIcon,
   o3fan: o3fanColorIcon,
   cherryin: cherryinColorIcon,
   modelscope: modelscopeColorIcon,